New configuration scheme is active
