--- Templates/gopher_sense_TEMPLATE.h.jinja2
@@ -1,25 +1,0 @@
-// Generated by SensorCannon
-#ifndef GOPHER_SENSE_H
-#define GOPHER_SENSE_H
-
-#include "sensor_hal.h"
-
-#define NUM_ANALOG_SENSORS {{analog_sensors|length}}
-#define NUM_CAN_SENSORS {{can_sensors|length}}
-
-
-{% for sensor in analog_sensors -%}
-extern ANALOG_SENSOR {{sensor.name}};
-{% endfor %}
-
-{% for sensor in can_sensors -%}
-extern CAN_SENSOR {{sensor.name}};
-{% endfor %}
-
-#define RES_OPEN -1.0
-
-void init_analog_sensors(void);
-void init_can_sensors(void);
-
-
-#endif // GOPHER_SENSE_H
