+++ Templates/hwconfig_TEMPLATE.c.jinja2
@@ -1,10 +1,9 @@
 // Autogenerated by SensorCannon
 // Hardware config for {{module.name}}
 
-#include "dam_hw_config.h"
+#include "module_hw_config.h"
 #include "gopher_sense.h"
 #include "GopherCAN.h"
-#include "DAM.h"
 
 // BUCKETS
 {% if buckets|length > 0 -%}
@@ -15,7 +14,7 @@ GENERAL_PARAMETER {{bucket.name}}_bucket_general_param_list[{{bucket.params|leng
 {
   {% for param in bucket.params -%}
   {
-    .can_param = (CAN_INFO_STRUCT*)(&{{param}}),
+    .can_param = (CAN_INFO_STRUCT*)(&{{param.param_name}}),
     .status = CLEAN,
 	.last_tx = (U32)(-1*MAX_TIME_BETWEEN_TX_ms)
   },
@@ -27,9 +26,10 @@ GENERAL_PARAMETER {{bucket.name}}_bucket_general_param_list[{{bucket.params|leng
 BUCKET bucket_list[NUM_BUCKETS] =
 {
   {% for bucket in buckets %}
+  {% set i = loop.index %}
   // Define bucket {{bucket.name}}
   {
-    .bucket_id = {{bucket.id}},
+    .bucket_id = {{i-1}},
     .ms_between_req = {{bucket.ms_between_req}},
     .state = BUCKET_CONFIG_INIT,
     .param_list = 
@@ -45,15 +45,15 @@ BUCKET bucket_list[NUM_BUCKETS] =
 {% endif %}
 // ***************************************************************************
 
-{% set l = module.adc1_params|length%}
-{% if l > 0 %}
+{% set len = module.adc1_params|length%}
+{% if len > 0 %}
 // BEGIN ADC1 PARAMS
 
 // give memory to the buffers
 {% for param in module.adc1_params %}
 {% set i = loop.index %}
-#define ADC1_PARAM{{i}}_BUF_SIZE {{param.buffer_size}}
-U32 adc1_param{{i}}_buffer_mem[ADC1_PARAM{{i}}_BUF_SIZE];
+#define ADC1_PARAM{{i}}_BUF_SIZE {{param.samples_buffered}}
+U16 adc1_param{{i}}_buffer_mem[ADC1_PARAM{{i}}_BUF_SIZE];
 {% endfor %}
 // Fill out the array with all of the sensor params
 ANALOG_SENSOR_PARAM adc1_sensor_params[NUM_ADC1_PARAMS] = 
@@ -63,9 +63,7 @@ ANALOG_SENSOR_PARAM adc1_sensor_params[NUM_ADC1_PARAMS] =
   // adc1_param{{i}}
   {
     .bucket_param = {{param.bucket_name}}_bucket_general_param_list + {{param.bucket_loc}},
-    .analog_sensor = &{{module.getSensorName(param.sensor_name)}},
-    .num_filtered_subparams = 0, // not yet implemented
-    .filtered_subparams = NULL, // not yet implemented
+    .analog_sensor = &{{param.param_name}},
     .buffer =
     {
       .buffer = adc1_param{{i}}_buffer_mem,
@@ -81,16 +79,15 @@ ANALOG_SENSOR_PARAM adc1_sensor_params[NUM_ADC1_PARAMS] =
 // No parameters on ADC1
 {% endif %}
 // ***************************************************************************
-
-{% set l = module.adc2_params|length%}
-{% if l > 0 %}
+{% set len = module.adc2_params|length%}
+{% if len > 0 %}
 // BEGIN ADC2 PARAMS
 
 // give memory to the buffers
 {% for param in module.adc2_params %}
 {% set i = loop.index %}
-#define ADC2_PARAM{{i}}_BUF_SIZE {{param.buffer_size}}
-U32 adc2_param{{i}}_buffer_mem[ADC2_PARAM{{i}}_BUF_SIZE];
+#define ADC2_PARAM{{i}}_BUF_SIZE {{param.samples_buffered}}
+U16 adc2_param{{i}}_buffer_mem[ADC2_PARAM{{i}}_BUF_SIZE];
 {% endfor %}
 // Fill out the array with all of the sensor params
 ANALOG_SENSOR_PARAM adc2_sensor_params[NUM_ADC2_PARAMS] = 
@@ -100,9 +97,7 @@ ANALOG_SENSOR_PARAM adc2_sensor_params[NUM_ADC2_PARAMS] =
   // adc2_param{{i}}
   {
     .bucket_param = {{param.bucket_name}}_bucket_general_param_list + {{param.bucket_loc}},
-    .analog_sensor = &{{module.getSensorName(param.sensor_name)}},
-    .num_filtered_subparams = 0, // not yet implemented
-    .filtered_subparams = NULL, // not yet implemented
+    .analog_sensor = &{{param.param_name}},
     .buffer =
     {
       .buffer = adc2_param{{i}}_buffer_mem,
@@ -118,16 +113,15 @@ ANALOG_SENSOR_PARAM adc2_sensor_params[NUM_ADC2_PARAMS] =
 // No parameters on ADC2
 {% endif %}
 // ***************************************************************************
-
-{% set l = module.adc3_params|length%}
-{% if l > 0 %}
+{% set len = module.adc3_params|length%}
+{% if len > 0 %}
 // BEGIN ADC3 PARAMS
 
 // give memory to the buffers
 {% for param in module.adc3_params %}
 {% set i = loop.index %}
-#define ADC3_PARAM{{i}}_BUF_SIZE {{param.buffer_size}}
-U32 adc3_param{{i}}_buffer_mem[ADC3_PARAM{{i}}_BUF_SIZE];
+#define ADC3_PARAM{{i}}_BUF_SIZE {{param.samples_buffered}}
+U16 adc3_param{{i}}_buffer_mem[ADC3_PARAM{{i}}_BUF_SIZE];
 {% endfor %}
 // Fill out the array with all of the sensor params
 ANALOG_SENSOR_PARAM adc3_sensor_params[NUM_ADC3_PARAMS] = 
@@ -137,9 +131,7 @@ ANALOG_SENSOR_PARAM adc3_sensor_params[NUM_ADC3_PARAMS] =
   // adc3_param{{i}}
   {
     .bucket_param = {{param.bucket_name}}_bucket_general_param_list + {{param.bucket_loc}},
-    .analog_sensor = &{{module.getSensorName(param.sensor_name)}},
-    .num_filtered_subparams = 0, // not yet implemented
-    .filtered_subparams = NULL, // not yet implemented
+    .analog_sensor = &{{param.param_name}},
     .buffer =
     {
       .buffer = adc3_param{{i}}_buffer_mem,
@@ -153,47 +145,6 @@ ANALOG_SENSOR_PARAM adc3_sensor_params[NUM_ADC3_PARAMS] =
 };
 {% else %}
 // No parameters on ADC3
-{% endif %}
-// ***************************************************************************
-
-{% set l = module.can_params|length%}
-{% if l > 0%}
-// BEGIN CAN PARAMS
-// give memory to the buffers
-{% for param in module.can_params %}
-{% set i = loop.index %}
-#define CAN_PARAM{{i}}_BUF_SIZE {{param.buffer_size}}
-U32 can_param{{i}}_buffer_mem[CAN_PARAM{{i}}_BUF_SIZE];
-{% endfor %}
-// Fill out the array with all of the sensor params
-CAN_SENSOR_PARAM can_sensor_params[NUM_CAN_SENSOR_PARAMS] =
-{
-  {% for param in module.can_params %}
-  {% set i = loop.index %}
-  {% set name = module.getSensorName(param.sensor_name) %}
-  // can_param{{i}}
-  {
-
-    .bucket_param = {{param.bucket_name}}_bucket_general_param_list + {{param.bucket_loc}},
-    .can_sensor = &{{name}},
-    .message_idx = {{module.getDependencyMessageIndex(name, param.sensor_output)}},
-    .num_filtered_subparams = 0, // not yet implemented
-    .filtered_subparams = NULL, // not yet implemented
-    .buffer =
-    {
-      .buffer = can_param{{i}}_buffer_mem,
-      .buffer_size = CAN_PARAM{{i}}_BUF_SIZE,
-      .fill_level = 0,
-	  .buffer_head = 0
-    },
-	.new_buf_data = FALSE
-
-  },
-  #undef CAN_PARAM{{i}}_BUF_SIZE
-  {% endfor %}
-};
-{% else %}
-// No CAN params
 {% endif %}
 // ***************************************************************************
