+++ Templates/hwconfig_TEMPLATE.h.jinja2
@@ -1,8 +1,8 @@
 // Autogenerated by SensorCannon
 //
 
-#ifndef DAM_HW_CONFIG_H
-#define DAM_HW_CONFIG_H
+#ifndef MODULE_HW_CONFIG_H
+#define MODULE_HW_CONFIG_H
 
 #include "sensor_hal.h"
 
@@ -10,7 +10,6 @@
 #define NUM_ADC1_PARAMS {{module.adc1_params | length}}
 #define NUM_ADC2_PARAMS {{module.adc2_params | length}}
 #define NUM_ADC3_PARAMS {{module.adc3_params | length}}
-#define NUM_CAN_SENSOR_PARAMS {{module.can_params | length}}
 #define NUM_BUCKETS {{buckets|length}}
 
 // analog params must be in channel order
@@ -26,13 +25,9 @@ extern ANALOG_SENSOR_PARAM adc2_sensor_params[NUM_ADC2_PARAMS];
 extern ANALOG_SENSOR_PARAM adc3_sensor_params[NUM_ADC3_PARAMS];
 #endif // NUM_ADC3_PARAMS > 0
 
-#if NUM_CAN_SENSOR_PARAMS > 0
-extern CAN_SENSOR_PARAM can_sensor_params[NUM_CAN_SENSOR_PARAMS];
-#endif // NUM_CAN_SENSOR_PARAMS > 0
-
 #if NUM_BUCKETS > 0
 extern BUCKET bucket_list[NUM_BUCKETS];
 #endif // NUM_BUCKETS > 0
 
-#endif // DAM_HW_CONFIG_H_H
+#endif // MODULE_HW_CONFIG_H
 // End autogenerated file
