+++ Templates/sensors_TEMPLATE.c.jinja2
@@ -1,0 +1,33 @@
+// Generated by SensorCannon
+
+#include "sensors.h"
+
+// ********** ANALOG_SENSORS **********
+{% for sensor in analog_sensors %}
+// Sensor definition: {{sensor.name_english}}
+{% set num = sensor.numTableEntries %}
+float {{sensor.name}}_table_independent_vars[{{num}}] = {
+    {% for entry in sensor.tableEntries[0] -%}
+    {{entry}}{% if loop.index < num %},{% endif %}
+    {% endfor %}
+};
+float {{sensor.name}}_table_dependent_vars[{{sensor.numTableEntries}}] = {
+    {% for entry in sensor.tableEntries[1] -%}
+    {{entry}}{% if loop.index < num %},{% endif %}
+    {% endfor %}
+};
+TABLE {{sensor.name}}_table = 
+{
+    .num_entries = {{num}},
+    .independent_vars = {{sensor.name}}_table_independent_vars,
+    .dependent_vars = {{sensor.name}}_table_dependent_vars
+};
+ANALOG_SENSOR {{sensor.name}} = 
+{
+    .type = {{sensor.analog_subtype.upper()}},
+	.conversion_table = {{sensor.name}}_table
+};
+
+{% endfor %}
+
+// END autogenerated file
