+++ Templates/sensors_TEMPLATE.h.jinja2
@@ -1,0 +1,19 @@
+// Generated by SensorCannon
+#ifndef SENSORS_H
+#define SENSORS_H
+
+#include "gsense_structs.h"
+
+#define NUM_ANALOG_SENSORS {{analog_sensors|length}}
+
+
+{% for sensor in analog_sensors -%}
+extern ANALOG_SENSOR {{sensor.name}};
+{% endfor %}
+
+
+void init_analog_sensors(void);
+void init_can_sensors(void);
+
+
+#endif // SENSORS_H
