--- sensor_hal.c
@@ -1,3 +1,0 @@
-#include "sensor_hal.h"
-
-
